+++ History.md
@@ -42,6 +42,7 @@
   * Update `Rack::Handler::Puma.run` to use `**options` (#2189)
   * ThreadPool concurrency refactoring (#2220)
   * JSON parse cluster worker stats instead of regex (#2124)
+  * Support parallel tests in verbose progress reporting (#2223)
 
 ## 4.3.3 and 3.12.4 / 2020-02-28